Inicio da aplicação dos padrões de projeto
